Fix speaking page bugs: nested anchors, duplicate component, ISO dates, and metadata type
- Extract ExternalLinkButton to shared component to eliminate duplication
- Fix nested anchor tags by restructuring card to only wrap image and title
- Add isoDate field to speaking data for valid JSON-LD structured data
- Add type: 'speaking' to detail page metadata for correct canonical URLs
- Add 'speaking' to TYPE_PATHS mapping in createMetadata utility
