+++ src/app/speaking/[slug]/page.jsx
@@ -3,6 +3,7 @@ import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { SimpleLayout } from '@/components/SimpleLayout'
 import { createMetadata } from '@/utils/createMetadata'
+import { ExternalLinkButton } from '@/components/ExternalLinkButton'
 import { speakingEngagements } from '../speaking-data'
 import { ArrowLeft, Calendar, Users, ExternalLink, Youtube, Link as LinkIcon, Play } from 'lucide-react'
 
@@ -21,6 +22,7 @@ export async function generateMetadata({ params }) {
     author: 'Zachary Proser',
     date: engagement.date,
     slug: `speaking/${slug}`,
+    type: 'speaking',
     image: engagement.image,
     keywords: engagement.topics,
     tags: ['speaking', ...engagement.topics],
@@ -31,42 +33,6 @@ export function generateStaticParams() {
   return speakingEngagements
     .filter(e => e.slug)
     .map(e => ({ slug: e.slug }))
-}
-
-function ExternalLinkButton({ link }) {
-  const getIcon = () => {
-    switch (link.type) {
-      case 'youtube': return <Youtube className="h-4 w-4" />
-      case 'blog': return <LinkIcon className="h-4 w-4" />
-      case 'twitter': return <ExternalLink className="h-4 w-4" />
-      default: return <ExternalLink className="h-4 w-4" />
-    }
-  }
-
-  const getButtonStyle = () => {
-    switch (link.type) {
-      case 'youtube': return 'bg-red-600 hover:bg-red-700 text-white'
-      case 'blog': return 'bg-burnt-400 hover:bg-burnt-500 dark:bg-amber-500 dark:hover:bg-amber-400 text-white'
-      case 'twitter': return 'bg-sky-500 hover:bg-sky-600 text-white'
-      default: return 'bg-gray-600 hover:bg-gray-700 text-white'
-    }
-  }
-
-  const isExternal = link.url.startsWith('http')
-  const LinkComponent = isExternal ? 'a' : Link
-  const linkProps = isExternal
-    ? { href: link.url, target: '_blank', rel: 'noopener noreferrer' }
-    : { href: link.url }
-
-  return (
-    <LinkComponent
-      {...linkProps}
-      className={`inline-flex items-center gap-2 px-3 py-2 rounded-md text-sm font-medium transition-colors ${getButtonStyle()}`}
-    >
-      {getIcon()}
-      {link.label}
-    </LinkComponent>
-  )
 }
 
 export default async function SpeakingDetail({ params }) {
@@ -196,7 +162,7 @@ export default async function SpeakingDetail({ params }) {
             '@type': 'Event',
             name: engagement.title,
             description: engagement.description,
-            startDate: engagement.date,
+            startDate: engagement.isoDate || engagement.date,
             location: {
               '@type': 'Place',
               name: engagement.location,
