+++ src/app/speaking/speaking-data.js
@@ -17,6 +17,7 @@ export const speakingEngagements = [
     title: 'Untethered Productivity: Staying Healthy, Creative, and Shipping in the AI Coding Era',
     event: 'AI Engineering London',
     date: 'April 2026',
+    isoDate: '2026-04-01',
     location: 'London, UK',
     description: 'A talk about the balance between leveraging AI coding agents for massive productivity gains while maintaining your health, creativity, and sanity. Covers signal management, context switching, developer balance, and building sustainable workflows with tools like Claude Code.',
     image: 'https://zackproser.b-cdn.net/talks/untethered-productivity/images/2026-04-05-hero-trail.png',
@@ -32,6 +33,7 @@ export const speakingEngagements = [
     title: 'Claude Cowork Workshop with Anthropic',
     event: 'WorkOS x Anthropic',
     date: 'February 26, 2026',
+    isoDate: '2026-02-26',
     location: 'San Francisco, CA',
     description: 'One-hour hands-on workshop I created and delivered at WorkOS, with Lydia from Anthropic\'s Claude Code team joining for Q&A. Demoed real-world Claude Code projects (Oura MCP, Handwave watchOS app, walking-and-talking development), then walked attendees through a complete GTM workflow: ICP identification, data scraping and enrichment, competitive analysis, battlecard creation, pain point messaging, cold email generation, blog content, and scheduled Cowork tasks for automated content production.',
     image: 'https://zackproser.b-cdn.net/images/workshop-zack-presenting-v2.webp',
@@ -63,6 +65,7 @@ export const speakingEngagements = [
     title: 'Keynote Speaker: DevSecCon 2025',
     event: 'DevSecCon 2025',
     date: '2025',
+    isoDate: '2025-01-01',
     location: 'Conference Keynote',
     description: 'Delivered the keynote address at DevSecCon 2025 on modern AI and security.',
     image: 'https://img.youtube.com/vi/kwIzRkzO_Z4/0.jpg',
@@ -83,6 +86,7 @@ export const speakingEngagements = [
     title: 'AI Pipelines and Agents in Pure TypeScript with Mastra.ai',
     event: 'AI Engineering World Fair',
     date: 'June 3, 2025',
+    isoDate: '2025-06-03',
     location: 'Workshop Session',
     description: 'We taught over 70 engineers over the course of a live 2 hour workshop how to build workflows to accomplish discrete tasks and how to grant access to those workflows to agents - which are the ideal human interface for accomplishing tasks with natural language.',
     image: aieWorkshop,
@@ -96,6 +100,7 @@ export const speakingEngagements = [
     title: 'Navigating from Jupyter Notebooks to Production',
     event: 'Pinecone & Cohere Meetup at a16z',
     date: 'December 6, 2023',
+    isoDate: '2023-12-06',
     location: 'Andreesen Horowitz, San Francisco, CA',
     description: 'I introduced the new Pinecone AWS Reference Architecture with Pulumi and explained infrastructure as code, using a mountaineering metaphor to compare getting from prototype to production.',
     image: a16z1,
@@ -121,6 +126,7 @@ export const speakingEngagements = [
     title: 'Voice-First Development & Claude Code Cowork Enablement',
     event: 'WorkOS GTM Training',
     date: 'February 6, 2026',
+    isoDate: '2026-02-06',
     location: 'WorkOS Office, San Francisco, CA',
     description: 'Demonstrated how to use WisprFlow to code with your voice and ran live demos of Claude Code and Cowork enablement for GTM teams. Co-presented with Nick Nisi.',
     image: 'https://zackproser.b-cdn.net/images/workos-gtm-training.webp',
@@ -141,6 +147,7 @@ export const speakingEngagements = [
     title: 'Claude Skills as Self-Documenting Runbooks',
     event: 'WorkOS Internal Training',
     date: 'October 23, 2025',
+    isoDate: '2025-10-23',
     location: 'WorkOS Office, San Francisco, CA',
     description: 'In-person training on operationalizing Claude Skills as executable, version-controlled runbooks and shareable team workflows.',
     image: claudeSkills,
@@ -161,6 +168,7 @@ export const speakingEngagements = [
     title: 'AI Fundamentals for Engineering Teams',
     event: 'WorkOS Internal Training',
     date: 'May 2025',
+    isoDate: '2025-05-01',
     location: 'In-person',
     description: 'Comprehensive introduction to AI concepts, machine learning fundamentals, and practical applications for software engineering teams. Covered LLMs, vector databases, RAG, and hands-on implementation strategies.',
     image: aiFundamentals,
@@ -174,6 +182,7 @@ export const speakingEngagements = [
     title: 'AI-Enabled Content Creation Workshop',
     event: 'WorkOS Internal Training',
     date: 'May 2025',
+    isoDate: '2025-05-01',
     location: 'In-person',
     description: 'Interactive workshop teaching content teams how to leverage AI tools for writing, editing, ideation, and content optimization. Practical hands-on session with real-world use cases and workflow optimization.',
     image: neuralNetworksLearn,
